type cast fix
Former-commit-id: 6373f5b30bd411461f76d2a6cf0afe32d131e3bc

+++ src/game/extinfo.h
@@ -97,7 +97,7 @@
         loopv(scores)
         {
             sendstring(teamtype[scores[i].team].name, p); //backward compatibility mode
-            putint(p, scores[i].score);
+            putint(p, (int)scores[i].score);
 
             if(m_stf(gamemode))
             {
